+++ src/Server/Services/Http/CrdControllers.cs
@@ -1,20 +1,21 @@
 ﻿/*
- * Apache License, Version 2.0
- * Copyright 2019-2020 NVIDIA Corporation
- * 
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *     http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+* Apache License, Version 2.0
+* Copyright 2019-2020 NVIDIA Corporation
+* 
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+* 
+*     http://www.apache.org/licenses/LICENSE-2.0
+* 
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*/
 
+using System;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
@@ -29,12 +30,13 @@ namespace Nvidia.Clara.DicomAdapter.Server.Services.Http
     public class ClaraAeTitleController : CrdCrudControllerBase<ClaraAeTitleController, ClaraApplicationEntity>
     {
         public ClaraAeTitleController(
+            IServiceProvider serviceProvider,
             IHttpContextAccessor httpContextAccessor,
             ILogger<ClaraAeTitleController> logger,
             IKubernetesWrapper kubernetesClient,
             ConfigurationValidator configurationValidator,
             IOptions<DicomAdapterConfiguration> dicomAdapterConfiguration)
-            : base(httpContextAccessor, logger, kubernetesClient, CustomResourceDefinition.ClaraAeTitleCrd, configurationValidator, dicomAdapterConfiguration)
+            : base(serviceProvider, httpContextAccessor, logger, kubernetesClient, CustomResourceDefinition.ClaraAeTitleCrd, configurationValidator, dicomAdapterConfiguration)
         {
         }
     }
@@ -44,12 +46,13 @@ namespace Nvidia.Clara.DicomAdapter.Server.Services.Http
     public class SourceAeTitleController : CrdCrudControllerBase<SourceAeTitleController, SourceApplicationEntity>
     {
         public SourceAeTitleController(
+            IServiceProvider serviceProvider,
             IHttpContextAccessor httpContextAccessor,
             ILogger<SourceAeTitleController> logger,
             IKubernetesWrapper kubernetesClient,
             ConfigurationValidator configurationValidator,
             IOptions<DicomAdapterConfiguration> dicomAdapterConfiguration)
-            : base(httpContextAccessor, logger, kubernetesClient, CustomResourceDefinition.SourceAeTitleCrd, configurationValidator, dicomAdapterConfiguration)
+            : base(serviceProvider, httpContextAccessor, logger, kubernetesClient, CustomResourceDefinition.SourceAeTitleCrd, configurationValidator, dicomAdapterConfiguration)
         {
         }
     }
@@ -59,12 +62,13 @@ namespace Nvidia.Clara.DicomAdapter.Server.Services.Http
     public class DestinationAeTitleController : CrdCrudControllerBase<DestinationAeTitleController, DestinationApplicationEntity>
     {
         public DestinationAeTitleController(
+            IServiceProvider serviceProvider,
             IHttpContextAccessor httpContextAccessor,
             ILogger<DestinationAeTitleController> logger,
             IKubernetesWrapper kubernetesClient,
             ConfigurationValidator configurationValidator,
             IOptions<DicomAdapterConfiguration> dicomAdapterConfiguration)
-            : base(httpContextAccessor, logger, kubernetesClient, CustomResourceDefinition.DestinationAeTitleCrd, configurationValidator, dicomAdapterConfiguration)
+            : base(serviceProvider, httpContextAccessor, logger, kubernetesClient, CustomResourceDefinition.DestinationAeTitleCrd, configurationValidator, dicomAdapterConfiguration)
         {
         }
     }